refactor-ir-cake: SigmaCompiler moved to sigma.compiler package
