+++ sc/shared/src/main/scala/org/ergoplatform/ErgoScriptPredef.scala
@@ -1,11 +1,11 @@
 package org.ergoplatform
 
-import sigmastate.lang.SigmaCompiler
 import org.ergoplatform.ErgoAddressEncoder.NetworkPrefix
 import sigma.ast.SType
 import sigma.ast.syntax.SigmaPropValue
 import sigma.ast.Value
 import sigma.ast.syntax.ValueOps
+import sigma.compiler.SigmaCompiler
 import sigma.compiler.ir.IRContext
 
 object ErgoScriptPredef {
