+++ sc/shared/src/main/scala/sigma/compiler/SigmaCompiler.scala
@@ -1,4 +1,4 @@
-package sigmastate.lang
+package sigma.compiler
 
 import fastparse.Parsed
 import fastparse.Parsed.Success
@@ -13,7 +13,9 @@ import sigma.ast._
 import sigma.ast.syntax.SValue
 import SCollectionMethods.{ExistsMethod, ForallMethod, MapMethod}
 import sigma.compiler.ir.{GraphIRReflection, IRContext}
+import sigma.compiler.phases.{SigmaBinder, SigmaTyper}
 import sigmastate.InterpreterReflection
+import sigmastate.lang.SigmaParser
 
 /**
   * @param networkPrefix    network prefix to decode an ergo address from string (PK op)
