+++ sc/shared/src/main/scala/sigma/compiler/SigmaTemplateCompiler.scala
@@ -1,10 +1,11 @@
-package sigmastate.lang
+package sigma.compiler
 
 import fastparse.Parsed
 import org.ergoplatform.sdk.{ContractTemplate, Parameter}
 import sigma.ast.SourceContext
 import sigma.ast.syntax.SValue
 import sigma.compiler.ir.CompiletimeIRContext
+import sigmastate.lang.{ContractParser, ParsedContractTemplate}
 import sigmastate.lang.parsers.ParserException
 
 /** Compiler which compiles Ergo contract templates into a [[ContractTemplate]]. */
