+++ sc/shared/src/test/scala/sigmastate/helpers/CompilerTestingCommons.scala
@@ -12,6 +12,7 @@ import sigma.data.{RType, SigmaBoolean}
 import sigma.validation.ValidationException
 import sigma.validation.ValidationRules.CheckSerializableTypeCode
 import sigma.ast.syntax.{SValue, SigmaPropValue}
+import sigma.compiler.{CompilerSettings, SigmaCompiler}
 import sigma.compiler.ir.IRContext
 import sigma.eval.{CostDetails, EvalSettings, Extensions, GivenCost, TracedCost}
 import sigmastate.helpers.TestingHelpers._
@@ -19,7 +20,6 @@ import sigma.interpreter.ContextExtension.VarBinding
 import sigmastate.interpreter.CErgoTreeEvaluator.DefaultProfiler
 import sigmastate.interpreter.Interpreter.ScriptEnv
 import sigmastate.interpreter._
-import sigmastate.lang.{CompilerSettings, SigmaCompiler}
 import sigma.serialization.SigmaSerializer
 import sigmastate.CompilerTestsBase
 import sigmastate.eval.CContext
